+++ rs/src/lib.rs
@@ -133,6 +133,7 @@ pub enum Tag {
   DefineGlyphFont(tags::DefineGlyphFont),
   DefineJpegTables(tags::DefineJpegTables),
   DefineMorphShape(tags::DefineMorphShape),
+  DefineScalingGrid(tags::DefineScalingGrid),
   DefineSceneAndFrameLabelData(tags::DefineSceneAndFrameLabelData),
   DefineShape(tags::DefineShape),
   DefineSound(tags::DefineSound),