+++ ts/src/lib/tag.ts
@@ -17,6 +17,7 @@ export type Tag =
   | tags.DefineGlyphFont
   | tags.DefineJpegTables
   | tags.DefineMorphShape
+  | tags.DefineScalingGrid
   | tags.DefineSceneAndFrameLabelData
   | tags.DefineShape
   | tags.DefineSound
@@ -66,6 +67,7 @@ export const $Tag: TaggedUnionType<Tag> = new TaggedUnionType<Tag>(() => ({
     tags.$DefineGlyphFont,
     tags.$DefineJpegTables,
     tags.$DefineMorphShape,
+    tags.$DefineScalingGrid,
     tags.$DefineSceneAndFrameLabelData,
     tags.$DefineShape,
     tags.$DefineSound,